+++ benchmarks/profiling.py
@@ -1,5 +1,8 @@
 """
 Sample script to profile the sql magic.
+
+>>> pip install line_profiler
+>>> kernprof -lv profiling.py
 """
 from sql.magic import SqlMagic
 from IPython import InteractiveShell
@@ -7,9 +10,10 @@ import duckdb
 from pandas import DataFrame
 import numpy as np
 
-num_rows = 1000_000
+num_rows = 1_000_000
+num_cols = 50
 
-df = DataFrame(np.random.randn(num_rows, 20))
+df = DataFrame(np.random.randn(num_rows, num_cols))
 
 magic = SqlMagic(InteractiveShell())
 
@@ -22,7 +26,7 @@ magic.displaycon = False
 # NOTE: you can put the @profile decorator on any internal function to profile it
 # the @profile decorator is injected by the line_profiler package at runtime, to learn
 # more, see: https://github.com/pyutils/line_profiler
-@profile  # noqa
+# e.g., to check the magic performance, you can add @profile to the _execute function
 def run_magic():
     magic.execute("SELECT * FROM df")